requirements.txt
pydantic~=2.11.5
httpx~=0.28.1
python-dotenv~=1.1.0